refactor(front): move assistant misc DTO types out of lib/api into types/api
Seventh batch of the lib/api DTO migration: assorted internal assistant
endpoints. Transport DTO types and zod schemas move to the mirrored path under
types/api/*, leaving business logic and domain types in place.

- 2 pure-contract files moved wholesale: builder/slack/channels_linked_with_agent,
  go_template_types.
- 9 files trimmed: agent_suggestion, agent_usage, mcp_configurations (back-imports
  AgentMcpConfigurationSummary used as a fn return type), messages, models,
  observability/summary, onboarding, rate_limits, tag_manager (moved the zod
  GetSuggestionsResponseBody; left io-ts WorkspaceTagSuggestionsResponseSchema).
- 23 consumers updated; 8 front-api routes split (DTO -> types, function -> lib);
  front/lib/swr/assistants.ts repointed 4 DTO imports.
