+++ front-api/routes/w/[wId]/assistant/agent_configurations/[aId]/observability/summary.ts
@@ -1,8 +1,8 @@
 import { DEFAULT_PERIOD_DAYS } from "@app/components/agent_builder/observability/constants";
 import { getAgentConfiguration } from "@app/lib/api/assistant/configuration/agent";
-import type { GetAgentSummaryResponseBody } from "@app/lib/api/assistant/observability/summary";
 import { generateAgentObservabilitySummary } from "@app/lib/api/assistant/observability/summary";
 import { buildAgentAnalyticsBaseQuery } from "@app/lib/api/assistant/observability/utils";
+import type { GetAgentSummaryResponseBody } from "@app/types/api/assistant/observability/summary";
 import { workspaceApp } from "@front-api/middlewares/ctx";
 import type { HandlerResult } from "@front-api/middlewares/utils";
 import { apiError } from "@front-api/middlewares/utils";
