+++ front-api/routes/w/[wId]/assistant/agent_configurations/[aId]/usage.ts
@@ -1,6 +1,6 @@
-import type { GetAgentUsageResponseBody } from "@app/lib/api/assistant/agent_usage";
 import { getAgentUsage } from "@app/lib/api/assistant/agent_usage";
 import { getAgentConfiguration } from "@app/lib/api/assistant/configuration/agent";
+import type { GetAgentUsageResponseBody } from "@app/types/api/assistant/agent_usage";
 import { workspaceApp } from "@front-api/middlewares/ctx";
 import type { HandlerResult } from "@front-api/middlewares/utils";
 import { apiError } from "@front-api/middlewares/utils";
