+++ front/hooks/useOnboardingConversation.ts
@@ -1,6 +1,6 @@
-import type { PostSendOnboardingResponseBody } from "@app/lib/api/assistant/onboarding";
 import { clientFetch } from "@app/lib/egress/client";
 import { useAppRouter, useSearchParam } from "@app/lib/platform";
+import type { PostSendOnboardingResponseBody } from "@app/types/api/assistant/onboarding";
 import { useCallback, useEffect, useRef } from "react";
 
 interface UseOnboardingConversationProps {
