+++ front/lib/api/assistant/mcp_configurations.ts
@@ -1,24 +1,7 @@
 import { AgentMCPServerConfigurationModel } from "@app/lib/models/agent/actions/mcp";
 import { AgentConfigurationModel } from "@app/lib/models/agent/agent";
+import type { AgentMcpConfigurationSummary } from "@app/types/api/assistant/mcp_configurations";
 import { Op } from "sequelize";
-import { z } from "zod";
-
-export const AgentMcpConfigurationSummarySchema = z.object({
-  sId: z.string(),
-  name: z.string().nullable(),
-});
-
-export type AgentMcpConfigurationSummary = z.infer<
-  typeof AgentMcpConfigurationSummarySchema
->;
-
-export const GetAgentMcpConfigurationsResponseBodySchema = z.object({
-  configurations: z.array(AgentMcpConfigurationSummarySchema),
-});
-
-export type GetAgentMcpConfigurationsResponseBody = z.infer<
-  typeof GetAgentMcpConfigurationsResponseBodySchema
->;
 
 export async function listAgentMcpConfigurationsForAgent(params: {
   workspaceId: number;
