+++ front/lib/api/assistant/messages.ts
@@ -18,7 +18,6 @@ import logger from "@app/logger/logger";
 import type { AgentMCPActionWithOutputType } from "@app/types/actions";
 import type { LightAgentConfigurationType } from "@app/types/assistant/agent";
 import type {
-  AgentMessageStatus,
   AgentMessageType,
   CompactionMessageType,
   LegacyLightMessageType,
@@ -40,7 +39,6 @@ import {
   toRichAgentMentionType,
   toRichUserMentionType,
 } from "@app/types/assistant/mentions";
-import type { SkillType } from "@app/types/assistant/skill_configuration";
 import type { ContentFragmentType } from "@app/types/content_fragment";
 import { isContentFragmentType } from "@app/types/content_fragment";
 import type { ModelId } from "@app/types/shared/model_id";
@@ -1059,39 +1057,3 @@ export async function fetchConversationMessages<V extends MessageVariant>(
         : never,
   });
 }
-
-export type PostMessagesResponseBody = {
-  message: UserMessageType;
-  contentFragments: ContentFragmentType[];
-  agentMessages: AgentMessageType[];
-};
-
-// TODO remove after monday 2025-12-01 (once everyone has likely reloaded their browser)
-export interface LegacyFetchConversationMessagesResponse {
-  hasMore: boolean;
-  lastValue: number | null;
-  messages: LegacyLightMessageType[];
-}
-
-export interface FetchConversationMessagesResponse {
-  hasMore: boolean;
-  lastValue: number | null;
-  messages: LightMessageType[];
-}
-
-export type FetchConversationMessageResponse = {
-  message: MessageType;
-};
-
-export type FetchConversationMessageResponseLight = {
-  message: LightMessageType;
-};
-
-export type FetchConversationMessageActionResponse = {
-  action: AgentMCPActionWithOutputType;
-  messageStatus: AgentMessageStatus;
-};
-
-export type GetAgentMessageSkillsResponseBody = {
-  skills: SkillType[];
-};
