+++ front/lib/api/assistant/observability/summary.ts
@@ -179,7 +179,3 @@ export async function generateAgentObservabilitySummary({
   }
   return new Ok({ summaryText });
 }
-
-export type GetAgentSummaryResponseBody = {
-  summaryText: string;
-};
