+++ front/lib/api/assistant/onboarding.ts
@@ -23,10 +23,6 @@ import { asDisplayName } from "@app/types/shared/utils/string_utils";
 
 import { createConversation, postUserMessage } from "./conversation";
 
-export type PostSendOnboardingResponseBody = {
-  conversationId: string | null;
-};
-
 function getOnboardingAvailableTools(): Array<{
   sId: string;
   name: string;
