+++ front/lib/swr/assistants.ts
@@ -1,12 +1,9 @@
 import { DEFAULT_PERIOD_DAYS } from "@app/components/agent_builder/observability/constants";
 import { useSendNotification } from "@app/hooks/useNotification";
-import type { GetAgentUsageResponseBody } from "@app/lib/api/assistant/agent_usage";
-import type { GetSlackChannelsLinkedWithAgentResponseBody } from "@app/lib/api/assistant/builder/slack/channels_linked_with_agent";
 import type {
   AgentMessageFeedbackType,
   AgentMessageFeedbackWithMetadataType,
 } from "@app/lib/api/assistant/feedback";
-import type { GetAgentMcpConfigurationsResponseBody } from "@app/lib/api/assistant/mcp_configurations";
 import type { GetContextOriginResponse } from "@app/lib/api/assistant/observability/context_origin";
 import type { GetDatasourceRetrievalResponse } from "@app/lib/api/assistant/observability/datasource_retrieval";
 import type { GetDatasourceRetrievalDocumentsResponse } from "@app/lib/api/assistant/observability/datasource_retrieval_documents";
@@ -18,7 +15,6 @@ import type {
 } from "@app/lib/api/assistant/observability/messages_metrics";
 import type { GetAgentOverviewResponseBody } from "@app/lib/api/assistant/observability/overview";
 import type { GetSkillExecutionResponse } from "@app/lib/api/assistant/observability/skill_execution";
-import type { GetAgentSummaryResponseBody } from "@app/lib/api/assistant/observability/summary";
 import type { GetToolExecutionResponse } from "@app/lib/api/assistant/observability/tool_execution";
 import type {
   GetToolLatencyResponse,
@@ -42,7 +38,11 @@ import {
   useSWRWithDefaults,
 } from "@app/lib/swr/swr";
 import { BROWSER_TIMEZONE } from "@app/lib/swr/workspaces";
+import type { GetAgentUsageResponseBody } from "@app/types/api/assistant/agent_usage";
+import type { GetSlackChannelsLinkedWithAgentResponseBody } from "@app/types/api/assistant/builder/slack/channels_linked_with_agent";
 import type { GetAgentConfigurationsResponseBody } from "@app/types/api/assistant/configuration";
+import type { GetAgentMcpConfigurationsResponseBody } from "@app/types/api/assistant/mcp_configurations";
+import type { GetAgentSummaryResponseBody } from "@app/types/api/assistant/observability/summary";
 import type {
   AgentConfigurationType,
   AgentsGetViewType,
