+++ front/types/api/assistant/agent_usage.ts
@@ -1,0 +1,9 @@
+import { AgentUsageSchema } from "@app/types/assistant/agent";
+import { z } from "zod";
+
+export const GetAgentUsageResponseBodySchema = z.object({
+  agentUsage: AgentUsageSchema.nullable(),
+});
+export type GetAgentUsageResponseBody = z.infer<
+  typeof GetAgentUsageResponseBodySchema
+>;
