+++ front/types/api/assistant/builder/slack/channels_linked_with_agent.ts
