+++ front/types/api/assistant/go_template_types.ts
