+++ front/types/api/assistant/models.ts
@@ -1,0 +1,5 @@
+import type { ModelConfigurationType } from "@app/types/assistant/models/types";
+
+export type GetEnabledModelsResponseType = {
+  models: ModelConfigurationType[];
+};
