+++ front/types/api/assistant/observability/summary.ts
@@ -1,0 +1,3 @@
+export type GetAgentSummaryResponseBody = {
+  summaryText: string;
+};
