+++ front/types/api/assistant/onboarding.ts
@@ -1,0 +1,3 @@
+export type PostSendOnboardingResponseBody = {
+  conversationId: string | null;
+};
