+++ front/types/api/assistant/rate_limits.ts
@@ -1,0 +1,4 @@
+export type GetTrialMessageUsageResponseType = {
+  count: number;
+  limit: number;
+};
