+++ front/types/api/assistant/tag_manager.ts
@@ -1,0 +1,16 @@
+import { z } from "zod";
+
+const GetSuggestionsResponseBodySchema = z.object({
+  suggestions: z
+    .array(
+      z.object({
+        name: z.string(),
+        agents: z.array(z.object({ sId: z.string(), name: z.string() })),
+      })
+    )
+    .nullish(),
+});
+
+export type GetSuggestionsResponseBody = z.infer<
+  typeof GetSuggestionsResponseBodySchema
+>;
